correction
correction cleaning workdone
gestion css et affichage comments
ajout caroussel en association

+++ cleaning_homepage.php
@@ -12,7 +12,7 @@ $iduser = $infoCleaner[0][0];
 //categorie de la personne admin,cleaning assoc.
 $usertype = $infoCleaner[0][3];
 //id comment .
-$idcomment= $infoCleaner[0][7];
+$idcomment = $infoCleaner[0][7];
 //j'appelle la date du jour
 date_default_timezone_set("Europe/Paris");
 $date = new DateTime('now');
@@ -35,19 +35,17 @@ if (isset($_POST['workdone'])) {
 }
 
 //pour annuler workdone
-if (isset($_POST['cancelWork'])) {
-
+// print ($hide);
+if (isset($_POST['cancelWork']) && isset($_POST['workdone'])) {
+  print('bobo');
   $sql = "DELETE FROM `interventions` WHERE id_user IN ($iduser) ORDER BY id_intervention DESC limit 1;";
   $db->prepExec($sql);
   $_SESSION['message1'] = null;
-  $_POST['workdone'] = !isset($_POST['workdone']);
+  $_POST['workdone'] = null;
   $_POST['timePeriod'] = '';
   header('location:cleaning_homepage.php');
 }
-//je donne une valeur a message car il n'existe pas avant de cliquer et cela genere une erreur lors du chargement de la page
-if (isset($_SESSION['message1'])) {
-  $_SESSION['message1'] = '';
-}
+
 
 // sur le select je choisi un temps de travail que je viens update a mon workdone
 //puis je creer un message qui indiqueque le travail est réalisé
@@ -68,7 +66,7 @@ if (isset($_POST['logout'])) {
 }
 
 //si la personne fait partie de l'equipe cleaning alors elle peux voir les commentaires selectionnés.
-if ($usertype == 'cleaning' ) {
+if ($usertype == 'cleaning') {
   $arrayComment = $db->getCommentsCleaning($usertype);
   $elementComment = $db->queryRequest($arrayComment);
   //recupere sql pour les réponses car order different
@@ -79,7 +77,6 @@ if ($usertype == 'cleaning' ) {
   $day = explode(' ', $dateComment[2]);
   $_SESSION['time_stamp'] = $day[0] . "/" . $dateComment[1] . "/" . $dateComment[0];
 }
-
 
 if (isset($_POST['reply'])) {
   $_SESSION['disapear'] = 1;
@@ -119,11 +116,6 @@ if (isset($_POST['cancelReply'])) {
   $_SESSION['disapear'] = 0;
 }
 
-//je donne une valeur a message car il n'existe pas avant de cliquer et cela genere une erreur lors du chargement de la page
-if (isset($_SESSION['message'])) {
-  $_SESSION['message1'] = '';
-}
-
 ?>
 
 <!DOCTYPE html>
@@ -133,7 +125,7 @@ if (isset($_SESSION['message'])) {
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
-  <!-- <link href="style.css" rel="stylesheet"> -->
+  <link href="style.css" rel="stylesheet">
   <title>Homepage_cleaning</title>
 </head>
 
@@ -161,7 +153,7 @@ if (isset($_SESSION['message'])) {
   </nav>
 
 
-  <section class="jumbotron jumbotron-fluid">
+  <section>
     <section class="row ">
       <article class=" col-sm-12 col-md-12 col-12 ">
         <h1 class=" pl-2 " id="bienvenu">Hello ! <?php print($_SESSION['first_name']); ?> <?php print($_SESSION['last_name']); ?></h1>
@@ -172,13 +164,14 @@ if (isset($_SESSION['message'])) {
             <div class="lead row ">
               <h3 class="display-8 text-light">Did you work today ?</h3>
               <div>
-                <button class="btn" style="background:#ecb21f; font-size:1em;margin-bottom:10px" type="submit" <?php if (!isset($_POST['cancel']) == '' && isset($_POST['workdone']) && isset($_POST['timePeriod']) == null || isset($_SESSION['message1']) != '') {
-                                                                                                                  if (isset($_POST['cancel']) == '') { ?> disabled <?php }
-                                                                                                                                                                } ?> onchange="this.form.submit()" name="workdone" id='workdone'>WORK DONE</button>
+                <button class="btn" style="background:#ecb21f; font-size:1em;margin-bottom:10px" type="submit" <?php if (isset($_POST['workdone']) && isset($_POST['timePeriod']) == null || isset($_POST['cancelWork']) == '') {
+                                                                                                                ?> disabled <?php }
+                                                                                                                            ?> onchange="this.form.submit()" name="workdone" id='workdone'>WORK DONE</button>
               </div>
               <div>
                 <?php if (isset($_POST['workdone']) && isset($_POST['timePeriod']) == null) { ?>
-                  <select name="timePeriod" onchange="this.form.submit()">
+                  <select name="timePeriod" onchange="this.form.submit()" required>
+                    <option value=''>how many hours ?</option>
                     <option value='30 min'>30 min</option>
                     <option value='1h30'>1h30</option>
                     <option value='2h'>2</option>
@@ -189,21 +182,17 @@ if (isset($_SESSION['message'])) {
                   </select>
                 <?php } ?>
               </div>
-              <?php if (isset($_POST['cancel']) == '' && isset($_SESSION['message1'])) { ?>
+              <?php if (isset($_POST['cancelWork']) == '' && isset($_POST['timePeriod'])) { ?>
                 <div class="alert alert-warning alert-dismissible fade show darker" role="alert">
-                  
                   <p class="lead"> <?php print("CONGRATULATION !") ?><br> <?php print("You've been working the " . $dateMessage); ?> </p>
                   <form method="post">
                     <button type="submit" style="background:#ecb21f; font-size:0.7em;margin-bottom:10px" name='cancelWork' id='cancel' class="btn" onchange="this.form.submit()">
                       <span style="color:black">CANCEL RECORD</span>
                     </button>
                   </form>
-
                 </div>
               <?php } ?>
-              <!-- ?php if (isset($_POST['workdone']) && isset($_POST['timePeriod']) != null || isset($_SESSION['message1']) != '') { ?>
-          <p class="lead"> ?php print("CONGRATULATION !") ?><br> ?php print("You've juste register your work for the " . $dateMessage); ?> </p>
-        ?php } ? -->
+
             </div>
           </div>
         </form>
@@ -212,98 +201,89 @@ if (isset($_SESSION['message'])) {
     </section>
     <hr>
     <section>
-
       <div class="container">
         <div class="row">
-          <div class="col-sm-5 col-md-6 col-12 pb-4">
+          <div class="col-12 col-sm-12 col-md-5 col-lg-5">
             <h1>Conversations</h1>
-            <?php 
-             foreach ($elementComment as $row) { 
-              $idComent = $row['id_comment']; 
+            <?php
+            foreach ($elementComment as $row) {
+              $idComent = $row['id_comment'];
               if ($row['id_comment'] != null && $row['comment_id'] == 0) { ?>
-              <form method="post">
-                <div class="darker mt-4 text-justify">
-                  <!-- //si on veut ajouter un avatar aux utilisateurs -->
-                  <img src="https://i.imgur.com/yTFUilP.jpg" alt="avatar" class="rounded-circle" width="40" height="40">
-                  <h4><?php print $row['first_name']; ?></h4>
-                  <p><?php print $row['description']; ?></p><br>
-                  <input type="hidden" name="parentDestinat" value="<?php print $row['destination']; ?>">
-                  <input type="hidden" name="parent_id" value="<?php print $row['id_comment']; ?>">
-                  <span>sent : <?php print $_SESSION['time_stamp']; ?></span><br>
-                  <button type="submit" style="background:#ecb21f; font-size:0.7em;margin-bottom:10px" name='reply' id='reply' class="btn" onchange="this.form.submit()">
-                    <span style="color:black">REPLY</span>
-                  </button>
-                </div>
-              </form>
-          </div>
-          <div>
-              <?php
-                  foreach ($elementResponse as $row) {
-                    if ($idComent == $row['comment_id']) { ?>
-                  <div class="darker mt-4 text-end response">
+                <form method="post">
+                  <div class="darker mt-4 text-justify">
                     <!-- //si on veut ajouter un avatar aux utilisateurs -->
                     <img src="https://i.imgur.com/yTFUilP.jpg" alt="avatar" class="rounded-circle" width="40" height="40">
                     <h4><?php print $row['first_name']; ?></h4>
                     <p><?php print $row['description']; ?></p><br>
+                    <input type="hidden" name="parentDestinat" value="<?php print $row['destination']; ?>">
+                    <input type="hidden" name="parent_id" value="<?php print $row['id_comment']; ?>">
                     <span>sent : <?php print $_SESSION['time_stamp']; ?></span><br>
-                  
+                    <button type="submit" style="background:#ecb21f; font-size:0.7em;margin-bottom:10px" name='reply' id='reply' class="btn" onchange="this.form.submit()">
+                      <span style="color:black">REPLY</span>
+                    </button>
                   </div>
-
-              <?php }
+                </form>
+                <div>
+                  <?php foreach ($elementResponse as $row) {
+                    if ($idComent == $row['comment_id']) { ?>
+                      <div class="darker mt-4 text-end response">
+                        <!-- //si on veut ajouter un avatar aux utilisateurs -->
+                        <img src="https://i.imgur.com/yTFUilP.jpg" alt="avatar" class="rounded-circle" width="40" height="40">
+                        <h4><?php print $row['first_name']; ?></h4>
+                        <p><?php print $row['description']; ?></p><br>
+                        <span>sent : <?php print $_SESSION['time_stamp']; ?></span><br>
+                      </div>
+                  <?php }
                   } ?>
-            <?php } ?>
+                <?php } ?>
+              <?php } ?>
+                </div>
+          </div>
+          <div class="col-12 col-sm-12 col-md-7 col-lg-7">
+            <form id="algin-form" method="post" <?php if ($_SESSION['disapear'] == 0) { ?>hidden <?php } ?>>
+              <div class="darker mt-4 text-justify">
+                <div class="form-group">
+                  <h4>Leave a message</h4>
+                  <textarea name="msgreply" maxlength="60" cols="30" rows="5" class="form-control text-light" style="background-color: black;"></textarea>
+                </div>
+                <div class="form-group">
+                  <button name="replybutton" type="submit" onchange="this.form.submit()" id="post" class="btn">Post Reply</button>
+                  <button type="submit" style="background:#ecb21f; font-size:0.7em;margin-bottom:10px; margin-top:10px" name='cancelReply' class="btn" onchange="this.form.submit()">CANCEL REPLY</button>
 
-        <?php }?> 
+                </div>
+              </div>
+
+            </form>
+            <form id="algin-form" method="post" <?php if ($_SESSION['disapear'] == 1) { ?>hidden <?php } ?>>
+              <div class="darker mt-4 text-justify">
+                <div class="form-group">
+                  <h4>Leave a message</h4>
+                  <textarea name="msg" maxlength="60" cols="30" rows="5" class="form-control text-light" style="background-color: black;"></textarea>
+                </div>
+                <label>Choose your recipient</label>
+                <div class="form-check text-light ">
+                  <input type="radio" class="  form-check-input" id="radio1" name="optradio" value="admin">Administrator
+                  <label class=" form-check-label" for="radio1"></label>
+                </div>
+                <div class="form-check text-light">
+                  <input type="radio" class="form-check-input" id="radio2" name="optradio" value="association">My association
+                  <label class="form-check-label" for="radio2"></label>
+                </div>
+                <div class="form-check text-light">
+                  <input type="radio" class="form-check-input" id="radio3" name="optradio" value="cleaning">Cleaning staff
+                  <label class="form-check-label" for="radio3"></label>
+                </div>
+
+                <div class="form-group">
+                  <button type="submit" onchange="this.form.submit()" id="post" name="postinfo" class="btn">Post Message</button>
+                </div>
+              </div>
+            </form>
+          </div>
         </div>
-        </div>
-      </div>
-
-        <div class="col-lg-4 col-md-5 col-sm-4 offset-md-1 offset-sm-1 col-12 mt-4">
-          <form id="algin-form" method="post" <?php if ($_SESSION['disapear'] == 0) { ?>hidden <?php } ?>>
-            <div class="darker mt-4 text-justify">
-              <div class="form-group">
-                <h4>Leave a message</h4>
-                <textarea name="msgreply" maxlength="60" cols="30" rows="5" class="form-control text-light" style="background-color: black;"></textarea>
-              </div>
-              <div class="form-group">
-                <button  name="replybutton" type="submit" onchange="this.form.submit()" id="post" class="btn">Post Reply</button>
-                <button type="submit" style="background:#ecb21f; font-size:0.7em;margin-bottom:10px; margin-top:10px" name='cancelReply' class="btn" onchange="this.form.submit()">CANCEL REPLY</button>
-
-              </div>
-            </div>
-          </form>
-
-          <form id="algin-form" method="post" <?php if ($_SESSION['disapear'] == 1) { ?>hidden <?php } ?>>
-            <div class="darker mt-4 text-justify">
-              <div class="form-group">
-                <h4>Leave a message</h4>
-                <textarea name="msg" maxlength="60" cols="30" rows="5" class="form-control text-light" style="background-color: black;"></textarea>
-              </div>
-              <label>Choose your recipient</label>
-              <div class="form-check text-light ">
-                <input type="radio" class="  form-check-input" id="radio1" name="optradio" value="admin">Administrator
-                <label class=" form-check-label" for="radio1"></label>
-              </div>
-              <div class="form-check text-light">
-                <input type="radio" class="form-check-input" id="radio2" name="optradio" value="association">My association
-                <label class="form-check-label" for="radio2"></label>
-              </div>
-              <div class="form-check text-light">
-                <input type="radio" class="form-check-input" id="radio3" name="optradio" value="cleaning">Cleaning staff
-                <label class="form-check-label" for="radio3"></label>
-              </div>
-
-              <div class="form-group">
-                <button type="submit" onchange="this.form.submit()" id="post" name="postinfo" class="btn">Post Message</button>
-              </div>
-            </div>
-          </form>
-        </div>
-      </div>
       </div>
     </section>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
-    <script type="text/javascript" src="script.js"></script>
+  </section>
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
+  <script type="text/javascript" src="script.js"></script>
 </body>
-
-</html>
